pyproject.toml
[project]
name = "rag-equipment-manuals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.64.1",
    "python-dotenv>=1.2.1",
    "raganything>=1.2.8",
    "streamlit>=1.28.0",
    "torch>=2.9.1",
]
